feat(map_and_intersection_selection): Reformating MapCanvas into two classes, and setting up the working intersection selection.
